+++ runtime/common/src/xcm_configs.rs
@@ -41,7 +41,7 @@ use frame_support::{
 };
 use sp_core::Get;
 use sp_io::hashing::blake2_256;
-use sp_runtime::traits::{Hash, SaturatedConversion, Saturating, Zero};
+use sp_runtime::traits::{SaturatedConversion, Saturating, Zero};
 use sp_std::{borrow::Borrow, prelude::*, result::Result};
 
 /// Base balance required for the XCM unit weight.
@@ -213,27 +213,4 @@ pub struct AssetRegistrarMetadata {
 	pub symbol: Vec<u8>,
 	pub decimals: u8,
 	pub is_frozen: bool,
-}
-
-// Our AssetType. For now we only handle Xcm Assets
-#[derive(Clone, Eq, Debug, PartialEq, Ord, PartialOrd, Encode, Decode, TypeInfo)]
-pub enum AssetType {
-	Xcm(MultiLocation),
-}
-impl Default for AssetType {
-	fn default() -> Self {
-		Self::Xcm(MultiLocation::here())
-	}
-}
-impl From<MultiLocation> for AssetType {
-	fn from(location: MultiLocation) -> Self {
-		Self::Xcm(location)
-	}
-}
-impl Into<Option<MultiLocation>> for AssetType {
-	fn into(self) -> Option<MultiLocation> {
-		match self {
-			Self::Xcm(location) => Some(location),
-		}
-	}
 }
