+++ runtime/pangolin/src/pallets/asset_manager.rs
@@ -21,6 +21,9 @@ use crate::*;
 // substrate
 use frame_support::{dispatch::GetDispatchInfo, pallet_prelude::*, transactional};
 use frame_system::EnsureNever;
+use sp_runtime::traits::Hash;
+// polkadot
+use xcm::latest::prelude::*;
 
 // We instruct how to register the Assets
 // In this case, we tell it to create an Asset in pallet-assets
@@ -91,6 +94,29 @@ frame_support::parameter_types! {
 	);
 }
 
+// Our AssetType. For now we only handle Xcm Assets
+#[derive(Clone, Eq, Debug, PartialEq, Ord, PartialOrd, Encode, Decode, TypeInfo)]
+pub enum AssetType {
+	Xcm(MultiLocation),
+}
+impl Default for AssetType {
+	fn default() -> Self {
+		Self::Xcm(MultiLocation::here())
+	}
+}
+impl From<MultiLocation> for AssetType {
+	fn from(location: MultiLocation) -> Self {
+		Self::Xcm(location)
+	}
+}
+impl Into<Option<MultiLocation>> for AssetType {
+	fn into(self) -> Option<MultiLocation> {
+		match self {
+			Self::Xcm(location) => Some(location),
+		}
+	}
+}
+
 // Implementation on how to retrieve the AssetId from an AssetType
 // We simply hash the AssetType and take the lowest 128 bits
 impl From<AssetType> for crate::AssetId {
@@ -116,7 +142,7 @@ impl pallet_asset_manager::Config for Runtime {
 	type Balance = Balance;
 	type Currency = Balances;
 	type ForeignAssetModifierOrigin = Root;
-	type ForeignAssetType = xcm_configs::AssetType;
+	type ForeignAssetType = AssetType;
 	type LocalAssetDeposit = ConstU128<0>;
 	type LocalAssetIdCreator = LocalAssetIdCreator;
 	type LocalAssetModifierOrigin = EnsureNever<AccountId>;
